chore: runner in the runner component
